refactor done between two ressource classes
