+++ UcmEmpire/src/main/java/models/biomes/BiomeForest.java
@@ -4,6 +4,7 @@ import models.Constantes;
 import models.Entity;
 import models.BoardPackage.Case;
 import models.BoardPackage.CaseSpecial;
+import models.ressources.RessourceName;
 
 import java.util.Random;
 
@@ -20,7 +21,7 @@ public class BiomeForest implements IBiomes {
         Random random = new Random();
 
         if (random.nextInt(Constantes.PROBA_TREE) == Constantes.PROBA_TREE) {
-            CaseSpecial<EnumRessourcersTemp> ressourceCase = new CaseSpecial<>(EnumRessourcersTemp.BOIS);
+            CaseSpecial<RessourceName> ressourceCase = new CaseSpecial<>(RessourceName.BOIS);
             return ressourceCase;
         } else {
             Case<Entity> entityCase = new Case<>(null, true, true);
