+++ UcmEmpire/src/main/java/models/biomes/BiomeMontain.java
@@ -4,6 +4,7 @@ import models.Constantes;
 import models.Entity;
 import models.BoardPackage.Case;
 import models.BoardPackage.CaseSpecial;
+import models.ressources.RessourceName;
 
 import java.util.Random;
 
@@ -19,7 +20,7 @@ public class BiomeMontain implements IBiomes {
         Random random = new Random();
 
         if (random.nextInt(Constantes.PROBA_STONE) == Constantes.PROBA_STONE) {
-            CaseSpecial<EnumRessourcersTemp> ressourceCase = new CaseSpecial<>(EnumRessourcersTemp.PIERRE);
+            CaseSpecial<RessourceName> ressourceCase = new CaseSpecial<>(RessourceName.PIERRE);
             return ressourceCase;
         } else {
