+++ UcmEmpire/src/main/java/models/ressources/RessourceName.java
@@ -1,7 +1,18 @@
 package models.ressources;
 
 public enum RessourceName {
-        PIERRE,
-        BOIS,
-        NOURRITURE
+
+        BOIS("bois"),
+        PIERRE("pierre"),
+        NOURRITURE("nourriture"); //TODO : quantity set to a number and not infinity
+
+        private String type;
+
+        RessourceName(String type) {
+                this.type = type;
+        }
+
+        public String getType() {
+                return type;
+        }
 }
